requirements.txt
asgiref==3.7.2
fpdf==1.7.2
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.10.0
django-filter==2.4.0
django-rest-swagger==2.2.0
